Extract QuestionPrompter from CreateScript
